+++ CHANGELOG.rst
@@ -23,8 +23,9 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
-*none*
 
+- Corrected a use-after-free when reusing an issuer or subject from an ``X509`` object after the underlying object has been mutated.
+  `#709 <https://github.com/pyca/pyopenssl/pull/709>`_
 
 ----